ci: use arch-tagged images for static build
For 32-bits arches, we use 64-bit arches in 32-bit mode. Docker by
default wants to pull the image for the native arch, so unless we take
care, the binaries will be built for the wrong arch.

Use the arch tagged images we build to make sure we get the correct image.