db, fix: make directory permission changes immediate
This removes the deferred directory permission changing. Stronger
guarantee of directories having proper permissions are now given
for package post-upgrade scripts. This also reduces one full
walk of all packages during install.

I believe the original reasoning for the deferred mechanism was
the possibility of memory overhead in tracking the packages for
each directory. Massif analysis shows that the directory structure
overhead is small compared to all the other data we read, so
lets do this now.