extract_v3: calculate xattr hash for audit
Calculate and write the xattr hash to installed database. This
will go away once installed database is migrated to new format.

Since the conversion happens on installation of the package,
the issue will get fixed the package is reinstalled (apk fix)
or upgraded to new version with apk including this commit.

fixes #10986