database: always use CSUM HEX for scriptdb tar entry
Always use CSUM in HEX for scriptdb tar entry.

Currently we use HEX for MD5 CSUM and base64 for SHA1 CSUM. This is
problematic for the scripts.tar as in base64 the "/" char is allowed.
This makes the tar think the file is placed in a dedicated directory
causing inconsistent files in scripts.tar making it problematic if
someone wants to handle scripts manually by accessing the file in the
.tar

Fix this by introducing apk_blob_push_csum_hex that always use HEX for
CSUM blob.

Pulling CSUM in HEX format is already supported and it's done by using
the X prefix instead of the current Q, hence this won't cause any
regression since both format are supported by the apk_blob_pull_csum
function.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>