tar: generalize apk_tar_write_padding size handling
Generalize apk_tar_write_padding size handling in preparation to make
use of it also in other context where apk_file_info is not used.

The function just required the size of data written hence we can just
pass that value directly instead of a pointer to the apk_file_info
struct.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>