mkpkg, extract_v3: make root directory optional in the PATHs
Since the root directory is generally not owned by any package
or even created by any package, it does not make sense to store
it in the packages unless needed (because it contains a file
in the root path).

Unfortunately, the extraction code assumed the first path is
the root path, so packages without the root node are not fully
compatible backwards.

Fix the extraction and make the code to omit the root node
behind an option. Eventually it will become the default and
the compatibility option will be removed.