apk_adb: remove ADB_VAL_NULL fallback from {oct,int}_fromstring
The return value of adb_w_int can never be 0 (cause the type is always
non-zero) which means the ADB_VAL_NULL fallback would never trigger.
If the intention was to return ADB_VAL_NULL when val is zero than it
would be better to handle that inside of the adb_w_int function.