change default db location to /usr, but detect and use /lib if exists
The current location of the DB under /lib is not FHS compliant, since
apk is not a boot-time component. Therefore, move the default location
to be placed under /usr, but keep using /lib for compatibility with
current installations. Downstreams are therefore responsible to move
the database location if they consider it a requirement.

The location is detected in the context and passed to other places
that might use it, as it is not only useful in the database