db: validate package arch against compatible arches
Allow the etc/apk/arch to be list of compatible arches. The first
one will be the primary arch which is used to construct index names.
The remaining arches are compatible arches that are accepted for
installation.

Validate packag arch to be a compatible arch to install it.

Always consider "noarch" to be a valid arch.

fixes #10844