mkpkg: fix creation of package with hardlinks
The initial scanning order is based on readdir() order. Before the
file data is processed, the ADB is sorted and the first hardlink
for a given file can change. If these orders mismatch, it would
cause an extraction failure.

To fix this, create the initial ADB with special place holders
for hardlink entries. After full scan, post-process the hardlink
targets so that the first sorted entry gets the data and subsequent
entries get the correct hardlink target name.

Add also a test case with multiple hardlinks.