build: use the VERSION directly as the soname
Currently no ABI guarantees are done so update soname on
every release.

Normally distribution adjust the display version to something
including the packaging version, but this avoids using that
currently.

It is left to the distribution to either build all reverse
dependencies on upgrade, or package libapk separately so
that multiple versions can coexist.

fixes #11001