db: make index refresh notices a progress note
The 'fetch ...' lines of update are not really applet output, but
status notifications to indicate that something network related
is happening.

Since that cannot go to stdout as some applets can output files
directly to stdout it was made to go to stderr. That also is
slightly problematic as it can be interpreted as an error, and
if both stdout/stderr are redirected to different files/pipes
their relative order might get jumbled.

Convert these notices to a new progress note that follows the
progress output scheme: erase the output once it is no longer
relevant. And follow the --(no-)progress state so these get
automatically disabled if output is piped.