mkndx: introduce --filter-spec to filter existing index
This can be used to create a subset of existing index.

E.g "abuild listpkg" could be used to create list of packages
of current repository and this list can be used to filter
existing index. Essentially removing packages from index
that have wrong version or which are not existent in repository
source.

Combined with --pkgname-spec with the original backing repository
URL, this can create an index with specific set of packages that
refer to upstream repository.

fixes #10873 for my use cases