commit: don't trigger question with --prune/--available warning
Don't make the --prune/--available note trigger the yes/no
question on whether to continue or not. This warning can appear
if no other changes are done on the system and the question does
not make sense in this case.

fixes 7b0b69a3 commit: advise on using upgrade --prune and/or --available