meson: Build Python module using standard Meson python build functions
This ensures that the Python module is built correctly against
the Python interpreter, uses the limited API automatically when
available, and installs into the correct path.

Unfortunately, this means that we are now using a feature in Meson
that does not exist yet in Muon for building the APK Python module.

That said, if you're already using Muon to build APK, you probably
do not want Python anyway.