apk: shared error message for invalid arguments
Previously only some invalid arguments printed an error message.

When getopt_long returned '?' it already printed an error message so
just exit.

This change also removes the usage after the error. The usage is long
which hides the error message in a wall of text. When inside of a tty
without scroll buffer that can prevent the user from ever reading the
error message.