db: do not remove installed files due to unexpected symlinks
In case of package upgrade, apk_db_purge_pkg() could remove newly
installed file if a file is moved between directories and there
is an unexpected symlink to make these package directories be
actually the same filesystem directory.

Track the device/inode of new files, and use this set to check
that the file to be removed is not something we just created.

fixes #11149