solver: fix swapping of inter depending packages via provides change
cset_gen_name_remove must not mark the name done as it will prevent
the processing of the name when the providing package name has
changed. This triggers when there's multiple package name changes
and they share common dependencies.

To keep the install-if removal correct, separate the name removed
state to a separate flag.