db: disable cache 'installed' on non-tmpfs installations
There are no real benefits for this, and may cause confusion
if the cache is shared between multiple hosts.

The intended usage if this has always been to store the
information about non-repository packages of run-from-tmpfs
installations.

Also, do not create the file if there is nothing to write,
and remove it if it become empty.

fixes #11138