ci: Run the test suite on Fedora Linux
With apk-tools being packaged in Fedora Linux and it being relied
on for kiwi, mkosi, and similar, it's good to make sure that apk
stays working on Fedora Linux.

The tests are also run in the packaging for Fedora Linux, and
running the test suite in CI should hopefully prevent surprises
in the future.