apk: refactor script output logging prefixes
The motivation to prefix script output with the script name, was
to provide greppable stdout. But this turns out to be fairly
cluttered and unreadable output.

Prefix script output with a properly indented "* " to indicate
its from script. And improve the commit hooks to be shown as
a temporary notice and generate an "executing hook" log line if
the script outputs something to know which hook generated
the output.

fixes #11159