db, pkg: add fallback to not needing unshare() when running as root
If running as root user, and --root is specified, add a fallback to
not depend on root. The scenario typically happens when running in
Docker and seccomp prevents unshare(). chroot() and mknod() still
works, so add back the creation of minimal /dev and disable memfd
since /proc is not mounted.

fixes #11166